+++ ChangeLog.md
@@ -1,6 +1,10 @@
 Change Log
 =
 
+2019-11-04:
+
+* update: add PreviewText config section
+
 2019-10-28:
 
 * update: add source update fetch control by `/set` command (usually used when source updating paused by reached 100 error when getting update data)